consolidated all feeder implementations in one file 'Feeder' documented file properly
